requirements.txt
# To install these Python dependencies, please type:
# pip install -r requirements.txt


streamlit>=0.62.0
numpy>=1.19.0
gensim>=3.8.3
pandas>=1.0.5
scikit-learn>=0.23.1
Pillow>=7.1.2
requests>=2.24.0
boto3>=1.14.7
PyYAML>=5.3.1
imblearn>=0.0
xgboost>=1.1.1